Updating Project to be able to submit as Group Project
We started off with looking at different data sets in New York and we decided with a data set that is 216,000 rows showing all of the business and individual licenses distributed throughout North America. We then decided that we wanted to figure out the statistics of license distribution in New York specifically. We decided to split it up by borough while also keeping a total of New York City. We then decided that we wanted to be more specific with our data so we made it show the amount of only the active licenses in each borough. Once we finished this we decided that we wanted to expand our project and wanted to calculate the ratio of the population of each borough to the amount of business licenses in each borough. To do this we found a different data set which we then took the population numbers of each borough and an estimate for the years ahead. After we finished that we learned how to make graphs so we decided to make two: a line graph that shows how the population changed over the past few years with the estimate for the next 20 years and a bar graph that demonstrates the number of businesses in each borough compared to each other. After we finished these graphs we decided to make a loop that allows the user to choose which graph they would like to see because we weren’t able to show both graphs in one go. So we created a while True loop which worked perfectly.